coinview/requirements.txt
python-binance
TA-Lib
numpy
backtrader
flask
matplotlib==3.2.2